requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
python-decouple>=3.8
whitenoise>=6.5.0
endesive>=2.0.14
cryptography>=41.0.0
pyOpenSSL>=23.0.0
requests>=2.31.0
